Added isSaveChanges flag to the repository

+++ src/BoltOn.Data.CosmosDb/Repository.cs
@@ -40,13 +40,17 @@ namespace BoltOn.Data.CosmosDb
             DocumentCollectionUri = UriFactory.CreateDocumentCollectionUri(DatabaseName, CollectionName);
         }
 
-        public virtual async Task<TEntity> AddAsync(TEntity entity, object options = null, CancellationToken cancellationToken = default)
+        public virtual async Task<TEntity> AddAsync(TEntity entity, object options = null, bool isSaveChanges = true, CancellationToken cancellationToken = default)
         {
-            PublishEvents(entity);
-            if (options is RequestOptions requestOptions)
-                await DocumentClient.CreateDocumentAsync(DocumentCollectionUri, entity, requestOptions, cancellationToken: cancellationToken);
-            else
-                await DocumentClient.CreateDocumentAsync(DocumentCollectionUri, entity, cancellationToken: cancellationToken);
+            if (isSaveChanges)
+            {
+                PublishEvents(entity);
+
+                if (options is RequestOptions requestOptions)
+                    await DocumentClient.CreateDocumentAsync(DocumentCollectionUri, entity, requestOptions, cancellationToken: cancellationToken);
+                else
+                    await DocumentClient.CreateDocumentAsync(DocumentCollectionUri, entity, cancellationToken: cancellationToken);
+            }
             return entity;
         }
 
@@ -99,23 +103,31 @@ namespace BoltOn.Data.CosmosDb
             }
         }
 
-        public virtual async Task UpdateAsync(TEntity entity, object options = null, CancellationToken cancellationToken = default)
+        public virtual async Task UpdateAsync(TEntity entity, object options = null, bool isSaveChanges = false, CancellationToken cancellationToken = default)
         {
-            PublishEvents(entity);
-            if (options is RequestOptions requestOptions)
-                await DocumentClient.UpsertDocumentAsync(DocumentCollectionUri, entity, requestOptions, cancellationToken: cancellationToken);
-            else
-                await DocumentClient.UpsertDocumentAsync(DocumentCollectionUri, entity, cancellationToken: cancellationToken);
+            if (isSaveChanges)
+            {
+                PublishEvents(entity);
+            
+                if (options is RequestOptions requestOptions)
+                    await DocumentClient.UpsertDocumentAsync(DocumentCollectionUri, entity, requestOptions, cancellationToken: cancellationToken);
+                else
+                    await DocumentClient.UpsertDocumentAsync(DocumentCollectionUri, entity, cancellationToken: cancellationToken);
+            }
 		}
 
-		public virtual async Task DeleteAsync(TEntity entity, object options = null, CancellationToken cancellationToken = default)
+		public virtual async Task DeleteAsync(TEntity entity, object options = null, bool isSaveChanges = false, CancellationToken cancellationToken = default)
 		{
-			PublishEvents(entity);
-			dynamic entityWithId = entity;
-			if (options is RequestOptions requestOptions)
-				await DocumentClient.DeleteDocumentAsync(GetDocumentUri(entityWithId.Id.ToString()), requestOptions);
-			else
-				await DocumentClient.DeleteDocumentAsync(GetDocumentUri(entityWithId.Id.ToString()));
+            if (isSaveChanges)
+            {
+                PublishEvents(entity);
+			    dynamic entityWithId = entity;
+
+                if (options is RequestOptions requestOptions)
+                    await DocumentClient.DeleteDocumentAsync(GetDocumentUri(entityWithId.Id.ToString()), requestOptions);
+                else
+                    await DocumentClient.DeleteDocumentAsync(GetDocumentUri(entityWithId.Id.ToString()));
+            }
 		}
 
 		protected Uri GetDocumentUri(string id)
